+++ WaterNut.Business.Services/Custom Services/DataModels/Custom DataModels/CreatingEx9/EX9Allocations.cs
@@ -32,7 +32,7 @@ namespace WaterNut.Business.Services.Custom_Services.DataModels.Custom_DataModel
         public string Customs_clearance_office_code { get; set; }
         public double? pQuantity { get; set; }
         public DateTime pRegistrationDate { get; set; }
-        public double Net_weight_itm { get; set; }
+        public decimal Net_weight_itm { get; set; }
         public double Total_CIF_itm { get; set; }
         public double Freight { get; set; }
         public double InternalFreight { get; set; }
